# openApi/swagger.yaml
openapi: 3.0.0
info:
  title: 'LaserLiga API'
  description: 'All API methods on the Laser liga portal.'
  version: '1.0'
servers:
  -
    url: 'https://laserliga.cz'
paths:
...
  '/api/leagues/{id}':
    get:
      tags:
        - Leagues
      summary: 'Get League by ID'
      description: 'This method returns a league based on the provided ID.'
      operationId: getLeague
      parameters:
        -
          name: id
          in: path
          description: 'League ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'League fetched successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/League'
        '403':
          description: 'Access denied'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/leagues/{id}/tournaments':
    get:
      tags:
        - Leagues
      summary: 'Get Tournaments of a League by ID'
      description: 'This method returns all the tournaments of a league based on the provided ID.'
      operationId: getLeagueTournaments
      parameters:
        -
          name: id
          in: path
          description: 'League ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Tournaments fetched successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tournament'
        '403':
          description: 'Access denied'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/leagues/{id}/points':
    get:
      tags:
        - Leagues
      summary: 'Recount points of a League by ID'
      description: 'This method recounts the points for each team in a league based on the provided league ID and returns the recalculated values.'
      operationId: recountLeaguePoints
      parameters:
        -
          name: id
          in: path
          description: 'League ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Points recounted successfully'
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items: { properties: { team: { type: string }, points: { type: integer }, positions: { type: object, additionalProperties: { properties: { position: { type: integer }, teamCount: { type: integer } }, type: object } } }, type: object }
        '403':
          description: 'Access denied'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/api/leagues/{id}/fixplayers':
    post:
      tags:
        - Leagues
      summary: 'Fix Players of a League by ID'
      description: 'This method fixes league players data based on the provided league ID and returns the result of the operation.'
      operationId: fixLeaguePlayers
      parameters:
        -
          name: id
          in: path
          description: 'League ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Players fixed successfully'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string }
                  players: { type: integer }
                  missing: { properties: { total: { type: integer }, foundMapPlayerCount: { type: integer }, foundPlayerCount: { type: integer }, newPlayerCount: { type: integer }, ids: { type: object } }, type: object }
                  teams: { type: object }
                type: object
        '500':
          description: 'Database error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Address:
      properties:
        street:
          type: string
          example: 'OD Luna 3. patro, Velké náměstí 175'
          nullable: true
        city:
          type: string
          example: Písek
          nullable: true
        postCode:
          type: string
          example: '39701'
          nullable: true
        country:
          type: string
          example: Česko
          nullable: true
      type: object
    Arena:
      properties:
        name:
          type: string
          example: 'Laser arena Písek'
        gameCodePrefix:
          type: string
          example: g
        lat:
          type: number
          format: float
          example: 49.307678
          nullable: true
        lng:
          type: number
          format: float
          example: 14.147773
          nullable: true
        address:
          $ref: '#/components/schemas/Address'
        web:
          type: string
          example: 'https://laserarenapisek.cz'
          nullable: true
        contactEmail:
          type: string
          example: info@laserarenapisek.cz
          nullable: true
        contactPhone:
          type: string
          example: '+420 776 606 631'
          nullable: true
      type: object
...
    EventPopup:
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        link:
          type: string
          nullable: true
        active:
          type: boolean
      type: object
...
    EventPriceGroup:
      properties:
        description:
          type: string
          nullable: true
        prices: {  }
      type: object
    League:
      properties:
        format: {  }
        teamLimit:
          type: integer
          nullable: true
        teamSize:
          type: integer
        subCount:
          type: integer
        registrationsActive:
          type: boolean
        registrationText:
          type: string
          nullable: true
        substituteRegistration:
          type: boolean
        requirements:
          $ref: '#/components/schemas/Requirements'
        name:
          type: string
        slug:
          type: string
          nullable: true
        shortDescription:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        popup:
          $ref: '#/components/schemas/EventPopup'
        registrationType:
          $ref: '#/components/schemas/RegistrationType'
        eventPriceGroup:
          oneOf:
            -
              $ref: '#/components/schemas/EventPriceGroup'
          nullable: true
        arena:
          $ref: '#/components/schemas/Arena'
      type: object
    RegistrationType:
      type: string
      enum:
        - league
        - tournament
        - both
    Requirement:
      type: string
      enum:
        - required
        - captain
        - optional
        - hidden
    Requirements:
      properties:
        playerName:
          $ref: '#/components/schemas/Requirement'
        playerSurname:
          $ref: '#/components/schemas/Requirement'
        playerEmail:
          $ref: '#/components/schemas/Requirement'
        playerParentEmail:
          $ref: '#/components/schemas/Requirement'
        playerPhone:
          $ref: '#/components/schemas/Requirement'
        playerParentPhone:
          $ref: '#/components/schemas/Requirement'
        playerBirthYear:
          $ref: '#/components/schemas/Requirement'
        playerSkill:
          $ref: '#/components/schemas/Requirement'
      type: object
    Tournament:
      properties:
        league:
          properties:
            id:
              type: integer
            name:
              type: string
          type: object
          nullable: true
        arena:
          $ref: '#/components/schemas/Arena'
        name:
          type: string
        eventPriceGroup:
          oneOf:
            -
              $ref: '#/components/schemas/EventPriceGroup'
          nullable: true
        rules:
          type: string
          nullable: true
        prices:
          type: string
          nullable: true
        resultsSummary:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        shortDescription:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        popup:
          $ref: '#/components/schemas/EventPopup'
        active:
          type: boolean
        finished:
          type: boolean
        format: {  }
        teamLimit:
          type: integer
          nullable: true
        teamSize:
          type: integer
        subCount:
          type: integer
        registrationsActive:
          type: boolean
        registrationText:
          type: string
          nullable: true
        substituteRegistration:
          type: boolean
        requirements:
          $ref: '#/components/schemas/Requirements'
        teamsInGame:
          type: integer
        points:
          $ref: '#/components/schemas/TournamentPoints'
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
          nullable: true
      type: object
    TournamentPoints:
      properties:
        win:
          type: integer
        draw:
          type: integer
        loss:
          type: integer
        second:
          type: integer
        third:
          type: integer
        other:
          type: array
          items:
            type: integer
      type: object
    ErrorResponse:
      properties:
        title:
          type: string
          example: 'Error title'
        type:
          $ref: '#/components/schemas/ErrorType'
        detail:
          type: string
          example: 'Error description'
          nullable: true
        exception:
          properties:
            message:
              type: string
              example: 'Some exception description'
            code:
              type: integer
              example: 123
            trace:
              type: array
              items:
                type: object
              example:
                -
                  file: index.php
                  line: 1
                  function: abc
                  args: ['Argument value']
          type: object
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
...
    ErrorType:
      type: string
      enum:
        - validation_error
        - database_error
        - internal_error
        - resource_not_found_error
        - resource_access_error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  -
    bearerAuth: []
...